backend_main/requirements.txt
coverage==7.8.0
django==5.1.6
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
itsdangerous==2.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
requests-mock==1.12.1
stripe==12.0.0
